% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccullagh.R
\name{mccullagh}
\alias{mccullagh}
\alias{dmccullagh}
\alias{pmccullagh}
\alias{qmccullagh}
\alias{rmccullagh}
\title{The McCullagh Distribution}
\source{
Pavia, J. (2015). GoFKernel: Testing Goodness-of-Fit with the Kernel Density Estimator.
}
\usage{
dmccullagh(x, theta = 0, nu = 4, log = FALSE)

pmccullagh(q, theta = 0, nu = 4, lower.tail = TRUE, log.p = FALSE)

qmccullagh(p, theta = 0, nu = 4, lower.tail = TRUE, log.p = FALSE)

rmccullagh(n, theta = 0, nu = 4)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{theta}{parameter. Must be between -1 and 1, inclusive.}

\item{nu}{parameter. Must be greater than -0.5.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\value{
\code{dmcullagh} gives the density, \code{pmccullagh} gives the distribution function, \code{qmccullagh} gives the quantile function, and \code{rmcculagh} generates random deviates.

The length of the result is determined by \code{n} for \code{rnorm}, and is the maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycled to the length of the result. Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile function and random generation for the McCullagh distribution with parameters \code{theta} and \code{nu}.
}
\details{
If \code{theta} or \code{nu} are not specified they assume the default values of 0 and 4, respectively.

The McCullagh distribution with parameters \eqn{\theta} (\code{theta}) and \eqn{\nu} (\code{nu}) has density

\deqn{f(x) = (1 - x^2)^(\nu - 0.5) / [ (1 - 2x\theta + \theta^2)^v * B(\nu + 0.5, 0.5) ]}

for -1 ≤ \eqn{x} ≤ 1, -1 < \eqn{\theta} < 1, \eqn{\nu} > - 0.5. The mean, variance, skewness and kurtosis are

\eqn{E(X) = \nu\theta / (\nu + 1)}

\eqn{Var(X) = (1 - [\nu * (\nu-1) * \theta^2] / [ (\nu + 1) * (\nu + 2) ] ) / [2 * (\nu + 1) ]}

\eqn{\kappa_3(X) = (\nu / [2 * (\nu + 1)^2 * (\nu + 2) ] ) * [ ( [ (\nu - 1) * (3\nu - 1) * \theta^3] / [ (\nu - 1) * (\nu + 3) ] ) - 3\theta ]}

\eqn{\kappa_4(X) = (-3 / [4 * (\nu + 1)^2 * (\nu + 2) ] ) * (1 - [4\nu * (3\nu - 1) * \theta^2] / [ (\nu + 1) * (\nu + 3) ] + ( [\nu (\nu - 1) * (11\nu^3 + 16\nu^2 - 17\nu + 2) * \theta^4] / [ (\nu + 1)^2 * (\nu + 2) * (\nu + 3) * (\nu + 4) ] ) )}
}
\note{
Because the cumulative distribution function Mccullagh does not have a closed form, the values of the functions \code{pmcullagh}, \code{qmcullagh} and \code{rmcullagh} are obtained through the solution of systems of nonlinear equations, which could delay the algorithm.

We will try to improve the times, especially in the function \code{rmccullagh}.
}
\examples{
dmccullagh(x = 0.3, theta = -0.8, nu = 2.3)

system.time(sim <- rmccullagh(n = 1000, theta = 0.2, nu = -0.15)) # it could take several seconds
plot(ecdf(sim))
curve(pmccullagh(q = x, theta = 0.2, nu = -0.15), add = TRUE, col = "red", lwd = 2)

qmccullagh(p = 0.2, theta = 0, nu = 0.41, lower.tail = FALSE)
}
\references{
McCullagh, P. (1989). Some statistical properties of a family of continuous univariate distributions. Journal of the American Statistical Association, 84(405), 125-129.

Seshadri, V. (1991). A family of distributions related to the McCullagh family. Statistics & probability letters, 12(5), 373-378.
}
\seealso{
\code{\link{Distributions}} for other standard distributions.
}
\author{
Jorge Iván Pérez, \email{jivan.perez@udea.edu.co}

Juan Carlos Correa, \email{jccorrea@unal.edu.co}
}
